+++ recover-config.sh
@@ -1,9 +1,13 @@
+#!/bin/bash
+
 if [[ ! "$PATH" =~ .*"solana/install/active_release/bin".* ]]; then
     gp env PATH="/home/gitpod/.local/share/solana/install/active_release/bin:$PATH"
 fi
 
 if [ -f "./.solana-config/id.json" ]; then
-    rm -rf /home/gitpod/.config/solana/*;
+    if [ -d "/home/gitpod/.config/solana" ]; then
+        rm -rf /home/gitpod/.config/solana/*;
+    fi
     cp -R /workspace/solana-helloworld/.solana-config/* /home/gitpod/.config/solana/;
 else 
     solana-keygen new --no-bip39-passphrase
